requirements.txt
runpod~=1.7.6

# Importante: fuerza wheels CUDA de PyTorch (cu121) dentro de la imagen
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchaudio

chatterbox-tts
